ref: css style, moved type files
