--- src/Common/Components/Filters/Filters.tsx
@@ -1,5 +1,0 @@
-const Filters = () => {
-  return <div>Filters</div>
-}
-
-export default Filters
