+++ src/Common/Models/TypeModels/RealtyAd.ts
