+++ src/Layouts/MainLayout.scss
@@ -5,6 +5,5 @@
 }
 
 .wrapper-outlet {
-  max-width: 50%;
-  padding: 20px;
+  padding: 20px 300px 20px 300px;
 }
