+++ src/Pages/AdPage/AdPage.scss
@@ -1,5 +1,20 @@
-.controls {
+.ad-wrapper {
+  width: 100%;
+  height: 100%;
   display: flex;
-  gap: 20px;
-  justify-content: space-between;
+  flex-direction: column;
+  gap: 15px;
+
+  .ad-list {
+    display: flex;
+    align-items: center;
+    gap: 30px;
+  }
+
+  .controls {
+    margin-top: 30px;
+    display: flex;
+    gap: 20px;
+    justify-content: space-between;
+  }
 }
