+++ src/Pages/FormPage/FormPage.scss
@@ -7,7 +7,16 @@
 
   .form-controls {
     display: flex;
-    justify-content: space-between;
+    gap: 12px;
+  }
+
+  .form-controls button {
+    flex-grow: 1;
+    max-width: 90%;
+
+    &:last-child {
+      width: 80px;
+    }
   }
 
   .steps-panel {
